+++ src/aux-server/aux-web/shared/vue-components/ShowAlertModal/ShowAlertModal.vue
@@ -1,0 +1,15 @@
+<template>
+    <div>
+        <md-dialog-alert
+            :md-active.sync="showAlertDialog"
+            :md-content="content"
+            :md-title="title"
+            :md-confirm-text="dismissText"
+            :md-fullscreen="true"
+            class="alert-dialog"
+            @md-confirm="onDismiss"
+        />
+    </div>
+</template>
+<script src="./ShowAlertModal.ts"></script>
+<style src="./ShowAlertModal.css" scoped></style>